refactor: consolidate into single executor, remove ExecuteSequential
Replace ExecuteSequential with Execute (parallel, per-layer concurrency).
Move Execute and changeToOp into executor.go, merge all tests into
executor_test.go (10 cases including transitive_skip and mixed_operations
ported from the old sequential suite). Delete executor_parallel.go and
executor_parallel_test.go.
